pyproject.toml
[project]
name = "pyglobegl"
version = "0.5.0"
description = "anywidget wrapper for globe.gl"
readme = "README.md"
authors = [
    { name = "Owen Lamont", email = "owenrlamont@gmail.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = [
    "anywidget",
    "globe.gl",
    "jupyter",
    "marimo",
    "webgl",
    "widget",
    "visualization",
]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
    "Typing :: Typed",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "anywidget>=0.9.21",
    "geojson-pydantic>=2.1.0",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "traitlets>=5.14.3",
]

[project.urls]
Repository = "https://github.com/owenlamont/pyglobegl"
Issues = "https://github.com/owenlamont/pyglobegl/issues"
Releases = "https://github.com/owenlamont/pyglobegl/releases"

[project.optional-dependencies]
geopandas = [
    "geopandas>=1.1.2",
    "pandera>=0.28.1",
    "webcolors",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "geopandas>=1.1.2",
    "ipykernel>=7.1.0",
    "jupyterlab>=4.5.1",
    "marimo>=0.19.0",
    "notebook>=7.5.1",
    "pandera>=0.28.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-ipywidgets[solara]>=1.56.0",
    "pytest-playwright>=0.7.2",
    "pytest-randomly>=4.0.1",
    "pytest-xdist>=3.8.0",
    "scikit-image>=0.25.2",
]
